Added key frame loop filtering support.
Note, the implemented loop filters can only be used on key frames.
Further modification is needed for inter frames, also the code is kinda
messy.
Also removed the -ftrapv compiler flags. During decoding only 16-bit
intergers are used, but CPU registers are 32-bit or 64-bit wide, so it's
unlikely for the CPU to report integer overflow.